Move mock window
