--- src/app/(site)/_components/mock-reviews.ts
@@ -1,22 +1,0 @@
-export const mockReviews = [
-  "Amazing service! They were quick, professional, and went above expectations. Highly recommend to anyone looking for quality work.",
-  "Best experience I've had in years. The attention to detail and customer care was outstanding. Will definitely be coming back!",
-  "Exceeded all my expectations. From start to finish, everything was handled perfectly. Five stars all the way!",
-  "Absolutely fantastic! The team was friendly, efficient, and delivered exactly what they promised. Couldn't be happier with the results.",
-  "Outstanding quality and service. They really know what they're doing and it shows in every detail. I'm impressed!",
-  "Super impressed with their work. Fast turnaround, great communication, and the final product was better than I imagined.",
-  "Incredible experience from beginning to end. Professional, courteous, and delivered beyond my expectations. Highly recommended!",
-  "Top-notch service all around. The staff was knowledgeable and helpful, making the whole process smooth and easy.",
-  "Brilliant work! They took the time to understand my needs and delivered a solution that was perfect. Will use again!",
-  "Very satisfied customer here. Great value, excellent quality, and wonderful customer service throughout. Thank you!",
-  "Fantastic job on everything. The attention to detail was impressive and they really went the extra mile for us.",
-  "Wonderful experience! They were patient, understanding, and delivered exactly what we needed. Couldn't ask for more!",
-  "Highly professional and skilled team. The results speak for themselves - absolutely beautiful work. Five stars deserved!",
-  "Exceptional service that truly stands out. They were responsive, creative, and delivered on time. Very happy customer!",
-  "Perfect from start to finish. Great communication, quality work, and fair pricing. What more could you want?",
-  "Really impressed with the level of care and expertise. They made everything so easy and the results are amazing!",
-  "Outstanding professionals who deliver what they promise. The quality exceeded my expectations and I'm thrilled with everything.",
-  "Absolutely love what they did! Creative, efficient, and professional throughout the entire process. Highly recommend!",
-  "Great experience working with them. They listened to my concerns and provided solutions that worked perfectly. Thank you!",
-  "Superb quality and attention to detail. The team was friendly and made sure everything was exactly right. Very pleased!",
-];
