+++ src/app/(site)/page.tsx
@@ -1,20 +1,17 @@
 "use client";
 
-import React, { useState } from "react";
+import React from "react";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
-import { Star, Code, Users, Globe, RefreshCw } from "lucide-react";
+import { Star, Code, Users, Globe } from "lucide-react";
 
 import { Avatar, AvatarImage } from "@/components/ui/avatar";
 import { Card } from "@/components/ui/card";
-import { MockWindow } from "@/components/ui/mock-window";
-import { Button } from "@/components/ui/button";
-import { Skeleton } from "@/components/ui/skeleton";
 
 import { ShimmerButton } from "@/components/magicui/shimmer-button";
 import PricingOptions from "@/components/common/pricing-options";
 
-import { mockReviews } from "./_components/mock-reviews";
+import { MockReviewsWindow } from "./_components/mock-reviews";
 
 const devCount = 24;
 const reviews = {
@@ -50,12 +47,6 @@ const frameworks = [
   { name: "11ty" },
 ];
 
-const exampleReviews = [
-  "Amazing service! They were quick, professional, and went above expectations. Highly recommend to anyone looking for quality work.",
-  "Best experience I've had in years. The attention to detail and customer care was outstanding. Will definitely be coming back!",
-  "Exceeded all my expectations. From start to finish, everything was handled perfectly. Five stars all the way!",
-];
-
 const steps = [
   {
     step: "01",
@@ -77,19 +68,6 @@ const steps = [
 
 export default function Home() {
   const router = useRouter();
-  const [displayedReviews, setDisplayedReviews] = useState(exampleReviews);
-  const [isLoading, setIsLoading] = useState(false);
-
-  const fetchRandomReviews = () => {
-    setIsLoading(true);
-
-    setTimeout(() => {
-      const shuffled = [...mockReviews].sort(() => 0.5 - Math.random());
-      const selected = shuffled.slice(0, 3);
-      setDisplayedReviews(selected);
-      setIsLoading(false);
-    }, 800);
-  };
 
   return (
     <>
@@ -109,82 +87,7 @@ export default function Home() {
           </div>
 
           {/* Product Screenshot */}
-          <MockWindow
-            title="ssg.tools - Google Reviews API"
-            theme="light"
-            className="shadow-2xl max-w-5xl mx-auto"
-          >
-            <div className="grid lg:grid-cols-2 gap-8 items-center p-2">
-              <div className="text-left">
-                <h3 className="text-lg font-semibold mb-4">API Endpoint</h3>
-                <div className="bg-gray-900 rounded-lg p-4 font-mono text-sm">
-                  <span className="text-green-400">GET</span>{" "}
-                  <span className="text-secondary">
-                    https://api.ssg.tools/reviews/
-                  </span>
-                  <span className="text-yellow-400">{"{business-id}"}</span>
-                </div>
-                <div className="flex content-end justify-end">
-                  <Button
-                    onClick={fetchRandomReviews}
-                    disabled={isLoading}
-                    className="mt-4 rounded-full"
-                    variant="default"
-                    size="sm"
-                  >
-                    Fetch
-                    <RefreshCw
-                      className={`w-3 h-3 ml-1 ${isLoading ? "animate-spin" : ""}`}
-                    />
-                  </Button>
-                </div>
-              </div>
-              <div className="text-left">
-                <h3 className="text-lg font-semibold mb-4">Fresh Reviews</h3>
-                <div className="space-y-3">
-                  {isLoading
-                    ? Array.from({ length: 3 }).map((_, i) => (
-                        <div
-                          key={i}
-                          className="flex items-center gap-3 p-3 bg-gray-50 rounded-lg h-[72px]"
-                        >
-                          <Skeleton className="w-10 h-10 rounded-full flex-shrink-0" />
-                          <div className="flex-1 space-y-2">
-                            <Skeleton className="h-4 w-24" />
-                            <Skeleton className="h-3 w-full" />
-                            <Skeleton className="h-3 w-4/5" />
-                          </div>
-                        </div>
-                      ))
-                    : displayedReviews.map((review, i) => (
-                        <div
-                          key={i}
-                          className="flex items-center gap-3 p-3 bg-gray-50 rounded-lg h-[72px]"
-                        >
-                          <div className="w-10 h-10 bg-secondary/20 rounded-full flex items-center justify-center flex-shrink-0">
-                            <span className="text-secondary font-semibold">
-                              {i + 1}
-                            </span>
-                          </div>
-                          <div className="flex-1 min-w-0">
-                            <div className="flex items-center gap-1 mb-1">
-                              {[...Array(5)].map((_, j) => (
-                                <Star
-                                  key={j}
-                                  className="w-4 h-4 fill-yellow-400 text-yellow-400"
-                                />
-                              ))}
-                            </div>
-                            <div className="text-sm text-gray-600 line-clamp-2">
-                              &quot;{review}&quot;
-                            </div>
-                          </div>
-                        </div>
-                      ))}
-                </div>
-              </div>
-            </div>
-          </MockWindow>
+          <MockReviewsWindow />
         </div>
       </section>
